move legacy code from tests/__init__.py to legacy/
